Fix explicit format for v7 in project format (#65)
* Make v7 the default version

Rojo v7.0 has been released and is fairly stable. It is therefore fair
to make it the default version used when exploring the Rojo docs.

* Fix explicit format for v7 in project format

Docs incorrectly reference the v6 way of using the explicit property
syntax rather than the new syntax found in v7 (described in the
`upgrade.md` file)

* Revert "Make v7 the default version"

This reverts commit 9a0a1c8d3ef5f17230b51f9f11458860cd5c8c7b.

* Change explicit format docs to be in v7 docs

The `v6` docs which is based on the master branch should not reflect
the explicit project format, only the `v7` docs under the
`versioned_docs` folder.